requirements.txt
# StatCastle requirements
pandas>=2.0
numpy>=1.24
requests>=2.31
beautifulsoup4>=4.12
pyyaml>=6.0
matplotlib>=3.8
reportlab>=4.0
jinja2>=3.1
playwright>=1.40
